+++ tests/integration/test_bbs_database.py
@@ -2,9 +2,10 @@ import subprocess
 import time
 
 import docker
+import pg8000
 import pytest
 import sqlalchemy
-from sqlalchemy.exc import OperationalError
+from sqlalchemy.exc import InterfaceError, OperationalError
 
 from bluesearch.entrypoint.database.parent import main
 
@@ -30,7 +31,8 @@ def get_docker_client():
     params=[
         "sqlite",
         # "mysql",  # not used in production and slows down CI
-        "mariadb",
+        # "mariadb", # not used in production and slows down CI
+        "postgres",
     ]
 )
 def setup_backend(request, tmp_path):
@@ -39,43 +41,86 @@ def setup_backend(request, tmp_path):
         db_url = tmp_path / "db.sqlite"
         yield "sqlite", str(db_url)
 
-    elif backend in {"mariadb", "mysql"}:
+    elif backend in {"mariadb", "mysql", "postgres"}:
+        # Check if docker client ready
         client = get_docker_client()
 
         if client is None:
             pytest.skip("Docker daemon is not running")
 
+        # Define backend specific logic
+        env_map = {
+            "mariadb": "MYSQL_ROOT_PASSWORD",
+            "mysql": "MYSQL_ROOT_PASSWORD",
+            "postgres": "POSTGRES_PASSWORD",
+        }
+
+        driver_map = {
+            "mariadb": "mysql+pymysql",
+            "mysql": "mysql+pymysql",
+            "postgres": "postgresql+pg8000",
+        }
+
+        port_map = {
+            "mariadb": 3306,
+            "mysql": 3306,
+            "postgres": 5432,
+        }
+
+        user_map = {
+            "mariadb": "root",
+            "mysql": "root",
+            "postgres": "postgres",
+        }
+
+        show_databases_map = {
+            "mariadb": "show databases",
+            "mysql": "show databases",
+            "postgres": "select datname from pg_database;",
+        }
+
+        create_database_map = {
+            "mariadb": "create database test",
+            "mysql": "create database test",
+            "postgres": "create database test;",
+        }
+
         port = 22346
+        pw = "my-secret-pw"
         container = client.containers.run(
             image=f"{backend}:latest",
-            environment={"MYSQL_ROOT_PASSWORD": "my-secret-pw"},
-            ports={"3306/tcp": port},
+            environment={f"{env_map[backend]}": pw},
+            ports={f"{port_map[backend]}/tcp": port},
             detach=True,
             auto_remove=True,
         )
 
         max_waiting_time = 2 * 60
         start = time.perf_counter()
-
+        url = f"{driver_map[backend]}://{user_map[backend]}:{pw}@127.0.0.1:{port}/"
         while time.perf_counter() - start < max_waiting_time:
             try:
-                engine = sqlalchemy.create_engine(
-                    f"mysql+pymysql://root:my-secret-pw@127.0.0.1:{port}/"
-                )
+                engine = sqlalchemy.create_engine(url, isolation_level="AUTOCOMMIT")
                 # Container ready?
-                engine.execute("show databases")
+                engine.execute(show_databases_map[backend]).fetchall()
                 break
-            except OperationalError:
+            except (
+                ConnectionResetError,
+                InterfaceError,
+                OperationalError,
+                pg8000.core.struct.error,
+                pg8000.dbapi.ProgrammingError,
+            ):
                 # Container not ready, pause and then try again
                 time.sleep(0.1)
                 continue
         else:
             raise TimeoutError("Could not spawn the container.")
 
-        engine.execute("create database test")
+        engine.execute(create_database_map[backend])
         engine.dispose()
 
-        yield backend, f"root:my-secret-pw@127.0.0.1:{port}/test",
+        yield backend, f"{user_map[backend]}:my-secret-pw@127.0.0.1:{port}/test",
 
         container.kill()
         client.close()
@@ -84,6 +129,7 @@ def setup_backend(request, tmp_path):
         raise ValueError
 
 
+@pytest.mark.filterwarnings("ignore::pytest.PytestUnraisableExceptionWarning")
 def test_bbs_database(tmp_path, setup_backend, jsons_path, caplog):
     # Parameters
     db_type, db_url = setup_backend
@@ -137,6 +183,9 @@ def test_bbs_database(tmp_path, setup_backend, jsons_path, caplog):
 
     elif db_type in {"mysql", "mariadb"}:
         engine = sqlalchemy.create_engine(f"mysql+pymysql://{db_url}")
+
+    elif db_type == "postgres":
+        engine = sqlalchemy.create_engine(f"postgresql+pg8000://{db_url}")
 
     query = "SELECT COUNT(*) FROM articles"
     (n_rows,) = engine.execute(query).fetchone()  # type: ignore